backend/dsp-mastering-engine/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.3
scipy==1.12.0
google-cloud-storage==2.14.0
pydantic==2.5.3
python-multipart==0.0.6
supabase

